Add effects on monsters for all magic traps
Previously most of the magic traps would have no effect when a monster
stepped on them. Now all magic traps will do something when triggered by
a monster.

+++ include/random.hpp
@@ -145,7 +145,7 @@ T element(const std::vector<T>& v)
 }
 
 template <typename T>
-size_t idx(const std::vector<T>& v)
+int idx(const std::vector<T>& v)
 {
         return range(0, v.size() - 1);
 }
